pyproject.toml
[project]
name = "type-speech"
version = "0.1.0"

dependencies = [
    "sounddevice==0.4.6",
    "numpy>=1.26.0",
    "google-cloud-speech==2.21.0",
    "pystray==0.19.5",
    "pyinstaller==6.14.2",
    "pillow==11.3.0",
    "loguru==0.7.2",
    "PyYAML>=6.0",
    "keyboard>=0.13.5",
    "pyperclip>=1.8.2",
    "pywin32>=306",
    "pydantic>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
build-exe = "build_exe:main"